Fake compressor duplication fixed
