+++ modules/vmfcore/test/test_compressor.cpp
@@ -156,42 +156,6 @@ TEST_P(TestCompressor, CheckRegisteredIds)
 }
 
 
-//Compressor class to pass fake Id to registerNew() method
-class FakeCompressor : public Compressor
-{
-public:
-    virtual void compress(const vmf_string& input, vmf_rawbuffer& output)
-    {
-        //copies input to another buffer and writes result
-        output = vmf_rawbuffer(input.c_str(), input.size());
-    }
-
-    virtual void decompress(const vmf_rawbuffer& input, vmf_string &output)
-    {
-        //copies code to another buffer and writes result
-        output = vmf_string(input.data(), input.size());
-    }
-
-    std::shared_ptr<Compressor> createNewInstance() const
-    {
-        return std::shared_ptr<Compressor>(new FakeCompressor);
-    }
-
-    virtual vmf::vmf_string getId()
-    {
-        return id;
-    }
-
-    void setId(vmf_string s)
-    {
-        id = s;
-    }
-
-private:
-    vmf_string id;
-};
-
-
 TEST_P(TestCompressor, TryRegisterExisting)
 {
     std::string name = GetParam();
